Move catalog search classes to Runs directory
They don't need to be in Common now they don't have a pointer to the main window

Re #23027
