+++ qt/scientific_interfaces/ISISReflectometry/GUI/CMakeLists.txt
@@ -43,7 +43,6 @@ set(GUI_MOC_FILES
     ${RUNS_MOC_FILES}
     ${RUNS_TABLE_MOC_FILES}
     ${SAVE_MOC_FILES}
-    ${COMMON_MOC_FILES}
     PARENT_SCOPE)
 
 set(GUI_UI_FILES
