+++ qt/scientific_interfaces/ISISReflectometry/GUI/Common/CMakeLists.txt
@@ -1,8 +1,6 @@
 set(
   COMMON_SRC_FILES
   Plotter.cpp
-  CatalogSearcher.cpp
-  SearchModel.cpp
 )
 
 # Include files aren't required, but this makes them appear in Visual Studio
@@ -12,9 +10,6 @@ set(
   IPythonRunner.h
   IPlotter.h
   Plotter.h
-  ISearcher.h
-  CatalogSearcher.h
-  SearchModel.h
 )
 
 set(COMMON_MOC_FILES SearchModel.h)
@@ -37,4 +32,3 @@ prepend(
 
 set(COMMON_SRC_FILES ${COMMON_SRC_FILES} PARENT_SCOPE)
 set(COMMON_INC_FILES ${COMMON_INC_FILES} PARENT_SCOPE)
-set(COMMON_MOC_FILES ${COMMON_MOC_FILES} PARENT_SCOPE)
