+++ qt/scientific_interfaces/ISISReflectometry/GUI/Runs/SearchResult.h
